+++ examples/ollama/CommunityToolkit.Aspire.Hosting.Ollama.AppHost/Program.cs
@@ -10,6 +10,7 @@ var llama = ollama.AddHuggingFaceModel("llama", "bartowski/Llama-3.2-1B-Instruct
 builder.AddProject<Projects.CommunityToolkit_Aspire_Hosting_Ollama_Web>("webfrontend")
     .WithExternalHttpEndpoints()
     .WithReference(phi3)
+    .WaitFor(phi3)
     .WithReference(llama);
 
 builder.Build().Run();
